test(runtime-config): split boundary guards into dedicated test file
